requirements.txt
asgiref==3.4.1
auditlog3==1.0.1
boto3==1.6.23
botocore==1.9.23
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.7
django-appconf==1.0.2
django-autocomplete-light==3.8.2
django-crispy-forms==1.12.0
django-filter==21.1
django-multiselectfield==0.1.10
django-safedelete==1.0.0
django-staticfiles==1.2.1
django-storages==1.6.6
docutils==0.14
gunicorn==19.7.1
jmespath==0.9.3
mysqlclient==2.0.3
Pillow==8.2.0
psycopg2-binary==2.8.6
python-dateutil==2.6.1
pytz==2018.3
s3transfer==0.1.13
six==1.16.0
sqlparse==0.4.1
static3==0.7.0
